requirements.txt
argh==0.26.2
pathtools==0.1.2
Pillow==6.1.0
PyYAML==5.1.2
watchdog==0.9.0
